degraded nltk to 3.0.4, fixed word2vec api server start error

+++ dependences/nltk/metrics/association.py
@@ -1,6 +1,6 @@
 # Natural Language Toolkit: Ngram Association Measures
 #
-# Copyright (C) 2001-2017 NLTK Project
+# Copyright (C) 2001-2015 NLTK Project
 # Author: Joel Nothman <jnothman@student.usyd.edu.au>
 # URL: <http://nltk.org>
 # For license information, see LICENSE.TXT
@@ -11,7 +11,6 @@ generic, abstract implementation in ``NgramAssocMeasures``, and n-specific
 ``BigramAssocMeasures`` and ``TrigramAssocMeasures``.
 """
 
-from __future__ import division
 import math as _math
 from functools import reduce
 _log2 = lambda x: _math.log(x, 2.0)
@@ -84,12 +83,12 @@ class NgramAssocMeasures(object):
             yield (_product(sum(cont[x] for x in range(2 ** cls._n)
                                 if (x & j) == (i & j))
                             for j in bits) /
-                   (n_all ** (cls._n - 1)))
+                   float(n_all ** (cls._n - 1)))
 
     @staticmethod
     def raw_freq(*marginals):
         """Scores ngrams by their frequency"""
-        return marginals[NGRAM] / marginals[TOTAL]
+        return float(marginals[NGRAM]) / marginals[TOTAL]
 
     @classmethod
     def student_t(cls, *marginals):
@@ -98,7 +97,7 @@ class NgramAssocMeasures(object):
         """
         return ((marginals[NGRAM] -
                   _product(marginals[UNIGRAMS]) /
-                  (marginals[TOTAL] ** (cls._n - 1))) /
+                  float(marginals[TOTAL] ** (cls._n - 1))) /
                 (marginals[NGRAM] + _SMALL) ** .5)
 
     @classmethod
@@ -118,7 +117,7 @@ class NgramAssocMeasures(object):
         logarithm of the result is calculated.
         """
         return (marginals[NGRAM] ** kwargs.get('power', 3) /
-                _product(marginals[UNIGRAMS]))
+                float(_product(marginals[UNIGRAMS])))
 
     @classmethod
     def pmi(cls, *marginals):
@@ -134,21 +133,21 @@ class NgramAssocMeasures(object):
         """
         cont = cls._contingency(*marginals)
         return (cls._n *
-                sum(obs * _ln(obs / (exp + _SMALL) + _SMALL)
+                sum(obs * _ln(float(obs) / (exp + _SMALL) + _SMALL)
                     for obs, exp in zip(cont, cls._expected_values(cont))))
 
     @classmethod
     def poisson_stirling(cls, *marginals):
         """Scores ngrams using the Poisson-Stirling measure."""
         exp = (_product(marginals[UNIGRAMS]) /
-               (marginals[TOTAL] ** (cls._n - 1)))
+               float(marginals[TOTAL] ** (cls._n - 1)))
         return marginals[NGRAM] * (_log2(marginals[NGRAM] / exp) - 1)
 
     @classmethod
     def jaccard(cls, *marginals):
         """Scores ngrams using the Jaccard index."""
         cont = cls._contingency(*marginals)
-        return cont[0] / sum(cont[:-1])
+        return float(cont[0]) / sum(cont[:-1])
 
 
 class BigramAssocMeasures(NgramAssocMeasures):
@@ -200,7 +199,7 @@ class BigramAssocMeasures(NgramAssocMeasures):
         n_xx = sum(cont)
         # For each contingency table cell
         for i in range(4):
-            yield (cont[i] + cont[i ^ 1]) * (cont[i] + cont[i ^ 2]) / n_xx
+            yield (cont[i] + cont[i ^ 1]) * (cont[i] + cont[i ^ 2]) / float(n_xx)
 
     @classmethod
     def phi_sq(cls, *marginals):
@@ -209,7 +208,7 @@ class BigramAssocMeasures(NgramAssocMeasures):
         """
         n_ii, n_io, n_oi, n_oo = cls._contingency(*marginals)
 
-        return ((n_ii*n_oo - n_io*n_oi)**2 /
+        return (float((n_ii*n_oo - n_io*n_oi)**2) /
                 ((n_ii + n_io) * (n_ii + n_oi) * (n_io + n_oo) * (n_oi + n_oo)))
 
     @classmethod
@@ -236,7 +235,7 @@ class BigramAssocMeasures(NgramAssocMeasures):
     def dice(n_ii, n_ix_xi_tuple, n_xx):
         """Scores bigrams using Dice's coefficient."""
         (n_ix, n_xi) = n_ix_xi_tuple
-        return 2 * n_ii / (n_ix + n_xi)
+        return 2 * float(n_ii) / (n_ix + n_xi)
 
 
 class TrigramAssocMeasures(NgramAssocMeasures):
